Move percent_change to Sanbase.Math and use it in more places
